Further refactor application
