+++ src/filekitty/core/python_parser.py
@@ -1,7 +1,7 @@
 import ast
 from pathlib import Path
 
-from .utils import read_file_contents
+from filekitty.core.utils import read_file_contents
 
 
 # --- Python Parsing Logic ---
