+++ src/filekitty/ui/qt_widgets.py
